Перенёс общую логику в NamedList
