+++ src/main/kotlin/Main.kt
@@ -1,7 +1,5 @@
-import java.util.Scanner
-
 fun main() {
-    val list1 = MenuItemList("Список архивов")
+    val list1 = ArchiveList("Список архивов")
  //   list1.add (MenuItemList("Архив1"))
  //   list1.add (MenuItemList("Архив2"))
  //   println(list1)
