+++ src/main/kotlin/NoteList.kt
@@ -1,0 +1,2 @@
+class NoteList {
+}
